modif for allowing repository delete to be int

+++ tatoulink/tatoulink/DataAccess/IRepository.cs
@@ -5,6 +5,6 @@
         Task<IEnumerable<ModelEntity>> Get(string includeTables = "");
         Task<ModelEntity> Insert(ModelEntity entity);
         Task<ModelEntity> Update(ModelEntity entity);
-        Task<bool> Delete(long idEntity);
+        Task<bool> Delete(int idEntity);
     }
 }
